+++ nginx_set_conf/__init__.py
@@ -5,7 +5,7 @@ nginx-set-conf - A tool for managing Nginx configurations
 from . import utils as utils
 from . import validators as validators
 
-__version__ = "1.17.0"
+__version__ = "1.17.1"
 
 __all__ = [
     "__version__",